added storage

+++ convex/core/storage/folder.ts
@@ -15,17 +15,19 @@ export const createFolder = mutation({
 export const deleteFolderById = mutation({
   args: {
     folderId: v.id('folders'),
+    userId: v.id('users'),
   },
   handler: async (ctx, args) => {
-    const { folderId } = args;
+    const { folderId, userId } = args;
     const folder = await ctx.db.get(folderId);
-    if (folder) {
+    const user = await ctx.db.get(userId);
+    if (folder && user) {
       const totalFilesLength = folder.fileLists.length;
 
       if (totalFilesLength > 0) {
         for (let i = 0; i <= totalFilesLength; i++) {
           const id = folder.fileLists[i] as Id<'files'>;
-          deleteFileById(ctx, { fileId: id });
+          deleteFileById(ctx, { fileId: id, userId });
         }
       }
       await ctx.db.delete(folderId);
